Fix publish release action as release-* branches are protected (#30)
Due to the branch protection, when this workflow is run in the
`release-v1.16.x` branch it fails due to the protection rules. And so,
no release is published.

Instead, the workflow checks that the `kubegres.yaml` is up to date with
the version requested to be released, if so it continues and creates the
tag and publishes the images, or else the workflow fails.

At the same time, this PR is updating and preparing for the next
release.