Active failback to standby (#41)
Part of https://github.com/tetrateio/tetrate/issues/25099

This PR:
- Makes the primary instance to be undeployed when
`spec.standby.enabled=false` is changed to `spec.standby.enabled=true`.
So, moving an active cluster to be a standby one.
- Improve the datetime logger messages to have a readable format when
running in k8s
- Labels the running instances with `kubegres.reactive-tech.io/mode` to
identify if they have been running in `active` or `standby` modes.
- Kills any replica running in active mode when it is changed back to
standby. This is done because an active replica can desynchronize with
the new primary instance while active and standby roles are swapped.
- Add tests