fix: standby replication slots primary endpoint
when we define primaryEndpoint in format $IP:$PORT a deploed replica in standby cluster fails.
without this replication slots cannot be created because of the way how we create connection to primary DB
that can be used by the operator.
this fix used predefined port as currently we don't allow to define port for primary endpoint

Signed-off-by: Piotr Kopec <piotr@tetrate.io>